Change Q3-c and the example in Q2-b
Q2-b: changed the example to the original example by Omer - showing its
possible for the test set to have lower MSE then the training set

Q3-c: change the RFE - we want to choose the most important features
based the whole Data. So with fit the RFE on X. After that we train the
model with the selected features on the data training set. At the end we
test our model with test set using the selected features.